requirements.txt
streamlit
google-genai
sentence-transformers
supabase
numpy